+++ pmd-scala-modules/pmd-scala-common/src/main/java/net/sourceforge/pmd/lang/scala/ast/ASTLitBoolean.java
@@ -16,7 +16,7 @@ public final class ASTLitBoolean extends AbstractScalaNode<Lit.Boolean> {
     }
 
     @Override
-    protected <P, R> R acceptVisitor(ScalaParserVisitor<? super P, ? extends R> visitor, P data) {
+    protected <P, R> R acceptVisitor(ScalaVisitor<? super P, ? extends R> visitor, P data) {
         return visitor.visit(this, data);
     }