+++ pmd-scala-modules/pmd-scala-common/src/main/java/net/sourceforge/pmd/lang/scala/rule/ScalaRule.java
@@ -7,13 +7,13 @@ package net.sourceforge.pmd.lang.scala.rule;
 import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.rule.AbstractRule;
-import net.sourceforge.pmd.lang.scala.ast.ScalaParserVisitor;
+import net.sourceforge.pmd.lang.scala.ast.ScalaVisitor;
 
 /**
  * The default base implementation of a PMD Rule for Scala. Uses the Visitor
  * Pattern to traverse the AST.
  */
-public class ScalaRule extends AbstractRule implements ScalaParserVisitor<RuleContext, RuleContext> {
+public class ScalaRule extends AbstractRule implements ScalaVisitor<RuleContext, RuleContext> {
 
     @Override
     public void apply(Node target, RuleContext ctx) {